crash: move Session into session module
We import everything required to setup a session just by importing
anything under crash.*.  That complicates testing.
